requirements.txt
fastapi
uvicorn
torch
transformers
huggingface_hub
pillow
python-multipart
